requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
flask-wtf==1.2.1
psycopg2-binary==2.9.9
anthropic>=0.40.0
celery==5.3.6
redis==5.0.1
stripe==7.8.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
werkzeug==3.0.1
wtforms==3.1.1
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
